+++ benchmarks/fuzzy_join_mini.py
@@ -7,233 +7,183 @@ of country names.
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
-from itertools import product
+from argparse import ArgumentParser
+from time import perf_counter
+import seaborn as sns
+import math
 
-from thefuzz.fuzz import partial_ratio, WRatio, ratio
+from thefuzz.fuzz import partial_ratio
 from thefuzz import process
 from autofj import AutoFJ
 from dirty_cat._fuzzy_join import fuzzy_join
+from utils import evaluate, fetch_data, default_parser, find_result, monitor
 
 
-def fetch_data(dataset_name):
-    """Fetch datasets from https://github.com/chu-data-lab/AutomaticFuzzyJoin/tree/master/src/autofj/benchmark
-    """
-    repository = "chu-data-lab/AutomaticFuzzyJoin"
-    base_url = f"https://raw.githubusercontent.com/{repository}/master/src/autofj/benchmark/{dataset_name}"  # noqa
-    left = pd.read_csv(f"{base_url}/left.csv")
-    right = pd.read_csv(f"{base_url}/right.csv")
-    gt = pd.read_csv(f"{base_url}/gt.csv")
-    return left, right, gt
+left_table, right_table, gt = fetch_data("Country")
 
 
-def thefuzz_merge(df_1, df_2, key1, key2, threshold=90, limit=2, scorer=partial_ratio):
+def thefuzz_merge(
+    df_1, df_2, left_on, right_on, threshold=0, limit=1, scorer=partial_ratio
+):
     """
     Merging using thefuzz
 
     Parameters:
         df_1: the left table to join
         df_2: the right table to join
-        key1: key column of the left table
-        key2: key column of the right table
+        left_on: key column of the left table
+        right_on: key column of the right table
         threshold: how close the matches should be to return a match, based on Levenshtein distance
         limit: the amount of matches that will get returned, these are sorted high to low
 
     Return:
-        dataframe with boths keys and matches
+        Dataframe with boths keys and matches.
     """
-    s = df_2[key2].tolist()
-
-    m = df_1[key1].apply(lambda x: process.extract(x, s, limit=limit, scorer=scorer))
+    s = df_2[right_on].tolist()
+    m = df_1[left_on].apply(lambda x: process.extract(x, s, limit=limit, scorer=scorer))
     df_1["matches"] = m
 
     m2 = df_1["matches"].apply(
         lambda x: ", ".join([i[0] for i in x if i[1] >= threshold])
     )
-
     df_1["matches"] = m2
+    thefuzz_join = df_1[df_1["matches"] != ""]
+    return thefuzz_join
 
-    return df_1[df_1["matches"] != ""]
 
-
-def test_autofj(left, right, gt, target):
+def autofj_merge(left, right, target=0.9):
     """Merging using AutomaticFuzzyJoin"""
     autofj = AutoFJ(precision_target=target, verbose=True)
-    LR_joins = autofj.join(left, right, id_column="id")
-
-    gt_joins = gt[["id_l", "id_r"]].values
-    LR_joins = LR_joins[["id_l", "id_r"]].values
-    p, r, f1 = evaluate(LR_joins, gt_joins)
-    print("Precision:", p, "Recall:", r, "F1:", f1)
-    return LR_joins, gt_joins
+    autofj_joins = autofj.join(left, right, id_column="id")
+    return autofj_joins
 
 
-def evaluate(pred_joins, gt_joins):
-    """Evaluate the performance of fuzzy joins
+#########################################################
+# Benchmarking accuracy and speed on actual datasets
+#########################################################
 
-    Parameters
-    ----------
-    pred_joins: list
-        A list of tuple pairs (id_l, id_r) that are predicted to be matches
-
-    gt_joins:
-        The ground truth matches
-
-    Returns
-    -------
-    precision: float
-        Precision score
-
-    recall: float
-        Recall score
-
-    f1: float
-        F1 score
-    """
-    pred = {(le, ri) for le, ri in pred_joins}
-    gt = {(le, ri) for le, ri in gt_joins}
-
-    tp = pred.intersection(gt)
-    precision = len(tp) / len(pred)
-    recall = len(tp) / len(gt)
-    # print('Precision', precision, 'Recall', recall)
-    if precision > 0 or recall > 0:
-        f1 = 2 * precision * recall / (precision + recall)
-    else:
-        f1 = 0
-    return precision, recall, f1
+benchmark_name = "fuzzy_join_full"
 
 
-def fuzzy_join_precision_recall(
-    left, right, gt, left_col, right_col, analyzer, ngram_range
+@monitor(
+    memory=True,
+    time=True,
+    parametrize={
+        "dataset_name": [
+            "Country",
+            # "BasketballTeam",
+            # "Drug",
+            # "Device",
+            # "ArtificialSatellite",
+            # "Amphibian",
+            # "Song",
+            # "HistoricBuilding",
+            # "Wrestler",
+            # "EthnicGroup",
+        ],
+        "join": [
+            "fuzzy_join",
+            "autofj",
+            "thefuzz",
+        ],
+    },
+    save_as=benchmark_name,
+    repeat=5,
+)
+def benchmark(
+    dataset_name: str,
+    join: str,
 ):
-    joined_fj = fuzzy_join(
-        right,
-        left,
-        left_on=left_col,
-        right_on=right_col,
-        analyzer=analyzer,
-        ngram_range=ngram_range,
-        return_score=True,
-    )
-    pr_list = []
-    re_list = []
-    f1_list = []
-    pr, re, f1 = evaluate(
-        list(zip(joined_fj["title_r"], joined_fj["title_l"])),
-        list(zip(gt["title_l"], gt["title_r"])),
-    )
-    pr_list.append(pr)
-    re_list.append(re)
-    f1_list.append(f1)
-    return pr_list, re_list, f1_list
+    left_table, right_table, gt = fetch_data(dataset_name)
 
-
-def thefuzz_precision_recall(
-    left, right, gt, left_col, right_col, scorer=partial_ratio
-):
-    pr_list = []
-    re_list = []
-    f1_list = []
-    for threshold in range(60, 99, 15):
-        joined = thefuzz_merge(
-            left,
-            right,
-            left_col,
-            right_col,
-            threshold=threshold,
-            limit=1,
-            scorer=scorer,
+    if join == "fuzzy_join":
+        start_time = perf_counter()
+        joined_fj = fuzzy_join(
+            left_table,
+            right_table,
+            how="left",
+            left_on="title",
+            right_on="title",
         )
+        end_time = perf_counter()
         pr, re, f1 = evaluate(
-            list(zip(joined[left_col], joined["matches"])),
+            list(zip(joined_fj["title_x"], joined_fj["title_y"])),
             list(zip(gt["title_l"], gt["title_r"])),
         )
-        pr_list.append(pr)
-        re_list.append(re)
-        f1_list.append(f1)
-    return pr_list, re_list, f1_list
+    elif join == "autofj":
+        start_time = perf_counter()
+        joined_fj = autofj_merge(
+            left_table,
+            right_table,
+        )
+        end_time = perf_counter()
+        pr, re, f1 = evaluate(
+            list(zip(joined_fj["title_l"], joined_fj["title_r"])),
+            list(zip(gt["title_l"], gt["title_r"])),
+        )
+    elif join == "thefuzz":
+        start_time = perf_counter()
+        joined_fj = thefuzz_merge(
+            left_table,
+            right_table,
+            left_on="title",
+            right_on="title",
+        )
+        end_time = perf_counter()
+
+    res_dic = {
+        "precision": pr,
+        "recall": re,
+        "f1": f1,
+        "time_fj": end_time - start_time,
+    }
+
+    return res_dic
 
 
-def autofj_precision_recall(left, right, gt, n_points=20):
-    pr_list = []
-    re_list = []
-    f1_list = []
-    for target in np.linspace(0.6, 1, n_points):
-        autofj = AutoFJ(precision_target=target, verbose=True)
-        LR_joins = autofj.join(left, right, id_column="id")
-        if len(LR_joins) != 0:
-            gt_joins = gt[["id_l", "id_r"]].values
-            LR_joins = LR_joins[["id_l", "id_r"]].values
-            p, r, f1 = evaluate(LR_joins, gt_joins)
-            pr_list.append(p)
-            re_list.append(r)
-            f1_list.append(f1)
-            # print("Precision:", p, "Recall:", r, "F1:", f1)
-    return pr_list, re_list, f1_list
+def plot(df: pd.DataFrame):
+    sns.set_theme(style="ticks", palette="pastel")
 
-
-def best_precision_recall(pr_list, re_list, n_bins=13):
-    bins = np.digitize(re_list, np.quantile(re_list, np.linspace(0, 1, n_bins)))
-    res_pr = np.zeros(n_bins - 1)
-    res_re = np.zeros(n_bins - 1)
-    for i in range(1, n_bins):
-        mask = bins == i
-        # print("Recall:", re_list[mask], "Precision:", pr_list[mask])
-        if len(pr_list[mask]) > 0:
-            res_pr[i - 1] = np.nanmax(pr_list[mask])
-            res_re[i - 1] = np.nanmean(re_list[mask])
-        else:
-            res_pr[i - 1] = np.nan
-            res_re[i - 1] = np.nan
-    return res_pr, res_re
+    n_datasets = len(np.unique(df["dataset_name"]))
+    n_rows = min(n_datasets, 3)
+    f, axes = plt.subplots(
+        n_rows,
+        math.ceil(n_datasets / n_rows),
+        squeeze=False,
+        figsize=(20, 5),
+    )
+    # Create the subplots but indexed by 1 value
+    for i, dataset_name in enumerate(np.unique(df["dataset_name"])):
+        sns.scatterplot(
+            x="time_fj",
+            y="f1",
+            hue="join",
+            # style="ngram_range",
+            # size="analyzer",
+            alpha=0.8,
+            data=df[df["dataset_name"] == dataset_name],
+            ax=axes[i % n_rows, i // n_rows],
+        )
+        axes[i % n_rows, i // n_rows].set_title(dataset_name)
+        # remove legend
+        axes[i % n_rows, i // n_rows].get_legend().remove()
+        # Put a legend to the right side if last row
+        if i == n_datasets - 1:
+            axes[i % n_rows, i // n_rows].legend(loc="center right")
+    plt.show()
 
 
 if __name__ == "__main__":
-    left_1, right_1, gt_1 = fetch_data("Country")
+    _args = ArgumentParser(
+        description="Benchmark for the batch feature of the MinHashEncoder.",
+        parents=[default_parser],
+    ).parse_args()
 
-    pr_list = []
-    re_list = []
-    for analyzer, max_n_gram in product(["char", "char_wb"], [3, 4, 5]):
-        if analyzer == "word" and max_n_gram > 2:
-            continue
-        precision, recall, f1 = fuzzy_join_precision_recall(
-            left=left_1,
-            right=right_1,
-            gt=gt_1,
-            left_col="title",
-            right_col="title",
-            analyzer=analyzer,
-            ngram_range=(2, max_n_gram),
-        )
-        pr_list.extend(precision)
-        re_list.extend(recall)
-        # plt.plot(recall, precision,
-        #         label=f"fuzzy_join_{analyzer}_{max_n_gram}_{similarity}")
+    if _args.run:
+        df = benchmark()
+    else:
+        result_file = find_result(benchmark_name)
+        df = pd.read_csv(result_file)
 
-    pr_list, re_list = best_precision_recall(np.array(pr_list), np.array(re_list))
-    plt.plot(re_list, pr_list, label="fuzzy_join")
-
-    pr_list_fw = []
-    re_list_fw = []
-    for scorer in partial_ratio, ratio, WRatio:
-        precision_fw, recall_fw, f1_fw = thefuzz_precision_recall(
-            left_1, right_1, gt_1, "title", "title", scorer=scorer
-        )
-        pr_list_fw.extend(precision_fw)
-        re_list_fw.extend(recall_fw)
-
-    pr_list_fw, re_list_fw = best_precision_recall(
-        np.array(pr_list_fw), np.array(re_list_fw)
-    )
-    plt.plot(re_list_fw, pr_list_fw, label="thefuzz")
-    # plt.plot(recall_fw, precision_fw, label=f'thefuzz_{scorer.__name__}')
-
-    precision_fj, recall_fj, f1_fj = autofj_precision_recall(left_1, right_1, gt_1)
-    plt.plot(recall_fj, precision_fj, label="autofj")
-
-    plt.xlabel("Recall")
-    plt.ylabel("Precision")
-    plt.legend()
-    plt.title("Best Precision / recall on Country")
-    plt.savefig("precision_recall.png")
-    # plt.show()
+    if _args.plot:
+        plot(df)
