+++ benchmarks/tuning_2023-05-08.log
@@ -1,0 +1,3 @@
+05/08/2023 12:30:03 - [INFO] Launching !
+05/08/2023 12:30:03 - [DEBUG] Creating pipelines
+05/08/2023 12:30:03 - [DEBUG] With cardinality possibilities: [20, 30, 40, 50] and n_components possibilities: [10, 30, 50]
