+++ benchmarks/utils/__init__.py
@@ -1,3 +1,3 @@
 from ._argparser import default_parser
-from ._various import choose_file, find_result, find_results
+from ._various import choose_file, find_result, find_results, evaluate, fetch_data
 from .monitor import monitor, repr_func
