+++ setup.cfg
@@ -58,6 +58,8 @@ benchmarks =
     matplotlib
     seaborn
     tqdm
+    thefuzz
+    autofj
 # Overwrite the previous install_requires for CI testing purposes
 # as defined in testing.yml.
 min-py38 =
